# doc/api.yaml
openapi: 3.0.0
info:
    title: Home Library Service
    description: Home music library service
    version: 1.0.0
servers:
    - url: /
components:
    schemas:
        User:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                login:
                    type: string
                    example: 'TestUser'
                version:
                    type: number
                    example: 1
                createAt:
                    type: number
                    example: 1655000000
                updateAt:
                    type: number
                    example: 1655000000
            required:
                - id
                - login
        Artist:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                name:
                    type: string
                    example: 'Freddie Mercury'
                grammy:
                    type: boolean
                    example: false
            required:
                - id
                - name
        Album:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                name:
                    type: string
                    example: 'Innuendo'
                year:
                    type: integer
                    example: 1991
                artistId:
                    type: string
                    format: uuid
                    nullable: true
            required:
                - id
                - name
...
        Login:
            type: object
            properties:
                userId:
                    type: string
                    format: uuid
                login:
                    type: string
                    example: 'TestUser'
                accessToken:
                    type: string
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI3MTY0N2RiMC1lYWI3LTExZWUtYWU4NC1lZmI2YmE2YTMxODMiLCJsb2dpbiI6ImFkbWluIiwiaWF0IjoxNzExMzc5MzQ5LCJleHAiOjE3MTEzODI5NDl9.YbQlf718qdZ_6yFO-Nz0aRqDgkh2-0FsVkKd1zv_fNc'
                refreshToken:
                    type: string
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI3MTY0N2RiMC1lYWI3LTExZWUtYWU4NC1lZmI2YmE2YTMxODMiLCJsb2dpbiI6ImFkbWluIiwiaWF0IjoxNzExMzc5MzQ5LCJleHAiOjE3MTE0NjU3NDl9.UT3nKF2VLa9J4IbSkrcGCYJ0mczeiYVsAbKUrkUSToU'
        Tokens:
            type: object
            properties:
                accessToken:
                    type: string
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI3MTY0N2RiMC1lYWI3LTExZWUtYWU4NC1lZmI2YmE2YTMxODMiLCJsb2dpbiI6ImFkbWluIiwiaWF0IjoxNzExMzc5MzQ5LCJleHAiOjE3MTEzODI5NDl9.YbQlf718qdZ_6yFO-Nz0aRqDgkh2-0FsVkKd1zv_fNc'
                refreshToken:
                    type: string
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI3MTY0N2RiMC1lYWI3LTExZWUtYWU4NC1lZmI2YmE2YTMxODMiLCJsb2dpbiI6ImFkbWluIiwiaWF0IjoxNzExMzc5MzQ5LCJleHAiOjE3MTE0NjU3NDl9.UT3nKF2VLa9J4IbSkrcGCYJ0mczeiYVsAbKUrkUSToU'
    responses:
        UnauthorizedError:
            description: Access token is missing or invalid
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
security:
    - bearerAuth: []
paths:
    /auth/login:
        post:
            tags:
                - Login
            security: []
            summary: Login
            description: Login a user and returns a JWT-tokens
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            title: example
                            properties:
                                login:
                                    type: string
                                    description: Username
                                password:
                                    type: string
                                    description: string
                            required:
                                - user
                                - login
            responses:
                200:
                    description: Successful login.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Login'
                400:
                    description: Bad request. Body does not contain required fields or they are not strings
                403:
                    description: Incorrect login or password. Doesn't match actual user
    /auth/signup:
        post:
            tags:
                - Signup
            security: []
            summary: Signup
            description: Signup a user
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            title: example
                            properties:
                                login:
                                    type: string
                                    minLength: 3
                                    maxLength: 255
                                    description: Username
                                password:
                                    type: string
                                    format: password
                                    pattern: ^[a-zA-Z0-9]{3,30}
                                    description: Password
                            required:
                                - login
                                - password
            responses:
                204:
                    description: Successful signup
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                400:
                    description: Bad request. Body does not contain required fields or they are not strings
    /auth/refresh:
        post:
            tags:
                - Refresh
            security: []
            summary: Refresh
            description: Refresh tokens. Get new pair of Access token and Refresh token
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            title: example
                            properties:
                                refreshToken:
                                    type: string
                            required:
                                - refreshToken
            responses:
                200:
                    description: Access token and Refresh token
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Tokens'
                401:
                    description: Bad request. Body does not contain required fields.
                403:
                    description: Authentication failed. Refresh token is invalid or expired.
...
    /album:
        get:
            tags:
                - Album
            summary: Get albums list
            description: Gets all library albums list
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Album'
                401:
                    $ref: '#/components/responses/UnauthorizedError'
        post:
            tags:
                - Album
            summary: Add new album
            description: Add new album information
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            title: example
                            properties:
                                name:
                                    type: string
                                year:
                                    type: integer
                                artistId:
                                    type: string
                                    format: uuid
                                    nullable: true
                            required:
                                - name
                                - year
            responses:
                201:
                    description: Album is created
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Album'
                400:
                    description: Bad request. body does not contain required fields
                401:
                    $ref: '#/components/responses/UnauthorizedError'
    /album/{id}:
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: string
                  format: uuid
        get:
            tags:
                - Album
            summary: Get single album by id
            description: Gets single album by id
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Album'
                400:
                    description: Bad request. albumId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Album was not found.
        put:
            tags:
                - Album
            summary: Update album information
            description: Update library album information by UUID
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            title: example
                            properties:
                                name:
                                    type: string
                                year:
                                    type: integer
                                artistId:
                                    type: string
                                    format: uuid
                                    nullable: true
                            required:
                                - name
                                - year
            responses:
                200:
                    description: The album has been updated.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Album'
                400:
                    description: Bad request. albumId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Album was not found.
        delete:
            tags:
                - Album
            summary: Delete album
            description: Delete album from library
            responses:
                204:
                    description: Deleted successfully
                400:
                    description: Bad request. albumId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Album was not found.
    /artist:
        get:
            tags:
                - Artist
            summary: Get all artists
            description: Gets all artists
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Artist'
                401:
                    $ref: '#/components/responses/UnauthorizedError'
        post:
            tags:
                - Artist
            summary: Add new artist
            description: Add new artist
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                grammy:
                                    type: boolean
                            required:
                                - name
                                - grammy
            responses:
                201:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Artist'
                400:
                    description: Bad request. body does not contain required fields
                401:
                    $ref: '#/components/responses/UnauthorizedError'
    /artist/{id}:
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: string
                  format: uuid
        get:
            tags:
                - Artist
            summary: Get single artist by id
            description: Get single artist by id
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Artist'
                400:
                    description: Bad request. artistId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Artist was not found.
        put:
            tags:
                - Artist
            summary: Update artist information
            description: Update artist information by UUID
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                grammy:
                                    type: boolean
                            required:
                                - name
                                - grammy
            responses:
                200:
                    description: The artist has been updated.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Artist'
                400:
                    description: Bad request. artistId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Artist was not found.
        delete:
            tags:
                - Artist
            summary: Delete artist
            description: Delete artist from library
            responses:
                204:
                    description: Deleted successfully
                400:
                    description: Bad request. artistId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Artist was not found.
...
    /favs/album/{id}:
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: string
                  format: uuid
        post:
            tags:
                - Favorites
            summary: Add album to the favorites
            description: Add album to the favorites
            responses:
                201:
                    description: Added successfully
                400:
                    description: Bad. albumId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                422:
                    description: Album with id doesn't exist.
        delete:
            tags:
                - Favorites
            summary: Delete album from favorites
            description: Delete album from favorites
            responses:
                204:
                    description: Deleted successfully
                400:
                    description: Bad. albumId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Album was not found.
    /favs/artist/{id}:
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: string
                  format: uuid
        post:
            tags:
                - Favorites
            summary: Add artist to the favorites
            description: Add artist to the favorites
            responses:
                201:
                    description: Added successfully
                400:
                    description: Bad request. artistId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                422:
                    description: Artist with id doesn't exist.
        delete:
            tags:
                - Favorites
            summary: Delete artist from favorites
            description: Delete artist from favorites
            responses:
                204:
                    description: Deleted successfully
                400:
                    description: Bad request. artistId is invalid (not uuid)
                401:
                    $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Artist was not found.